secure-wipe/python_app/requirements.txt
pyqt6==6.5.0
psutil==5.9.0
cryptography==41.0.0
reportlab==4.0.4
qrcode[pil]==7.4.2
flask==2.3.3
requests-oauthlib==1.3.1
pyyaml==6.0.1
pydantic==2.5.0